fixed invalid type assertion error

+++ metrics/internal/workers/system.go
@@ -81,24 +81,24 @@ func (s *system) Run(ctx context.Context, workersAmount int) {
 						continue
 					}
 
-					dur, ok := m.Data["time"].(int)
+					dur, ok := m.Data["time"].(float64)
 					if !ok {
 						continue
 					}
 
-					s.timePerRequest.With(prometheus.Labels{"req_name": reqName}).Observe(float64(dur))
+					s.timePerRequest.With(prometheus.Labels{"req_name": reqName}).Observe(dur)
 				case RequestStatus:
 					reqName, ok := m.Data["reqName"].(string)
 					if !ok {
 						continue
 					}
 
-					status, ok := m.Data["status"].(int)
+					status, ok := m.Data["status"].(float64)
 					if !ok {
 						continue
 					}
 
-					s.requestStatus.With(prometheus.Labels{"req_name": reqName, "status": strconv.Itoa(status)}).Inc()
+					s.requestStatus.With(prometheus.Labels{"req_name": reqName, "status": strconv.Itoa(int(status))}).Inc()
 				default:
 					l.Error("invalid message type", log.WithData("type", m.Type))
 				}
